upgrade Chess::Core::Pieces::Pawn
